login split into LoginForm
