+++ client/src/Components/LoginForm/LoginForm.css
@@ -1,0 +1,32 @@
+.login_btn {
+			width: 100%;
+			background: #c0392b !important;
+			color: white !important;
+		}
+		.login_btn:focus {
+			box-shadow: none !important;
+			outline: 0px !important;
+		}
+		.login_container {
+			padding: 0 2rem;
+		}
+		.input-group-text {
+			background: #c0392b !important;
+			color: white !important;
+			border: 0 !important;
+			border-radius: 0.25rem 0 0 0.25rem !important;
+            height: 38px;
+		}
+		.input_user,
+		.input_pass:focus {
+			box-shadow: none !important;
+			outline: 0px !important;
+		}
+        .fa-key,
+        .fa-user {
+	        padding-top: 11;
+	        padding-bottom: 11;
+        }
+        .links {
+            margin-top: 8px;
+        }
